requirements.txt
streamlit
vosk
gtts
groq
pydub
sounddevice
scipy
python-dotenv
